Removes SCF_PERMANENT * The SCF_PERMANENT flag was equivalent to SCF_NO_CLEARBUFF. * Adjusts script command sc_end to abide by this flag now.

+++ src/map/chrif.cpp
@@ -1647,7 +1647,7 @@ int chrif_bsdata_save(struct map_session_data *sd, bool quit) {
 
 	// Removing...
 	if (quit && sd->bonus_script.head) {
-		int64 flag = BSF_REM_ON_LOGOUT; //Remove bonus when logout
+		uint32 flag = BSF_REM_ON_LOGOUT; //Remove bonus when logout
 
 		if (battle_config.debuff_on_logout&1) //Remove negative buffs
 			flag |= BSF_REM_DEBUFF;
